Changing to be compatible with versions of chef older than 12.5
